:tada: Upgrade Ansible Role for TeleIRC v2.0.0
This is a first pass at updating the Ansible Role for the forthcoming
release of TeleIRC.

Note this commit also changes the license of the project. This moves it
from a Mozilla Public License 2.0 project to a BSD-3-Clause project.
Why? I realized a permissive license makes more sense to better allow
creative freedom on derivative works that use this Ansible Role. A
copyleft license does not encourage reuse in this specific context.

Signed-off-by: Justin W. Flory <git@jwf.io>